some code cleanup
